requirements.txt
fastapi
uvicorn
deta
python-decouple
python-multipart